save state

+++ app/src/main/java/com/example/price/SearchInfo.kt
@@ -1,7 +1,21 @@
 package com.example.price
 
-public class SearchInfo (var id : Int , var partName : String , var carName : String ,var creationTime : String , var carModel : String , var carChassis : String , var brandName : String ,
-                         var country : String , var price : Int , var supplierId : String , var stockCount : Int ,var description: String,var lastModificationTime : String)
-{
+public class SearchInfo(
+    var id: Int,
+    var partName: String,
+    var carName: String,
+    var creationTime: String,
+    var carModel: String,
+    var carChassis: String,
+    var brandName: String,
+    var country: String,
+    var price: Double,
+    var supplierId: String,
+    var stockCount: Int,
+    var description: String,
+    var lastModificationTime: String,
+
+
+) {
 
 }
